+++ apps/app/src/app/@buyer/feed/FeedListPage/FeedList/CreateButton.tsx
@@ -1,19 +1,8 @@
 import { Icon, PlusIcon } from "@use-pico/client/icon";
-import { BottomSheet } from "@use-pico/client/ui/bottom-sheet";
-import { Container } from "@use-pico/client/ui/container";
-import { FormField } from "@use-pico/client/ui/form";
-import { Mx } from "@use-pico/client/ui/mx";
-import { Status } from "@use-pico/client/ui/status";
-import { TextInput } from "@use-pico/client/ui/text-input";
 import { Tx } from "@use-pico/client/ui/tx";
-import { translator } from "@use-pico/common/translator";
-import { sFeedCreate } from "@zbav-se.me/sdk/api/buyer";
-import { withFeedQuery } from "@zbav-se.me/sdk/query/buyer/feed";
-import { CloseButton } from "@zbav-se.me/ui/button";
 import { type FC, useState } from "react";
-import { SaveContainer } from "~/app/@common/container/ui/SaveContainer";
-import { getFeedDefaultCreate } from "~/app/@common/feed/service/getFeedDefaultCreate";
 import { ListItem } from "~/app/@common/list-item/ListItem";
+import { CreateSheet } from "./CreateSheet";
 
 export namespace CreateButton {
 	export interface Props extends ListItem.PropsEx {
@@ -23,20 +12,6 @@ export namespace CreateButton {
 
 export const CreateButton: FC<CreateButton.Props> = ({ ui, className, ...props }) => {
 	const [isOpen, setIsOpen] = useState(false);
-	const [name, setName] = useState("");
-
-	const feedCreateMutation = withFeedQuery.useCreateMutation({
-		onSettled() {
-			setIsOpen(false);
-			setName("");
-		},
-		invalidate: [
-			"collection",
-			"count",
-		],
-	});
-
-	const invalid = !name || name.length < sFeedCreate.properties.name.minLength;
 
 	return (
 		<>
@@ -73,75 +48,12 @@ export const CreateButton: FC<CreateButton.Props> = ({ ui, className, ...props }
 				{...props}
 			/>
 
-			<BottomSheet
-				isOpen={isOpen}
-				onClose={() => setIsOpen(false)}
-				detent={"default"}
-				header={({ close }) => ({
-					title: "Create new feed (title)",
-					right: <CloseButton onClick={close} />,
-				})}
-			>
-				<Container
-					data-ui={"CreateButton[TextInputContainer]"}
-					ui={{
-						layout: "vertical-content-footer",
-						height: "full",
-						width: "full",
-						inner: "default",
-					}}
-				>
-					<Container
-						ui={{
-							layout: "vertical-centered",
-							height: "full",
-						}}
-					>
-						<Status
-							textTitle={translator.text("Feed name (title)")}
-							action={
-								<FormField>
-									{(props) => (
-										<TextInput
-											value={name}
-											onChange={(e) => {
-												setName(e.target.value);
-											}}
-											placeholder={translator.text("Feed name (placeholder)")}
-											autoFocus
-											minLength={sFeedCreate.properties.name.minLength}
-											{...props}
-										/>
-									)}
-								</FormField>
-							}
-							ui={{
-								text: "md",
-								inner: "4xl",
-							}}
-						>
-							<Mx
-								label={translator.text("Feed name (required)")}
-								ui={{
-									tone: "neutral",
-									theme: "light",
-								}}
-							/>
-						</Status>
-					</Container>
-
-					<SaveContainer
-						onCancel={() => {
-							setIsOpen(false);
-						}}
-						onSave={() => {
-							feedCreateMutation.mutate(getFeedDefaultCreate(name));
-						}}
-						loading={feedCreateMutation.isPending}
-						disabled={invalid || feedCreateMutation.isPending}
-					/>
-				</Container>
-			</BottomSheet>
+			<CreateSheet
+				state={{
+					value: isOpen,
+					set: setIsOpen,
+				}}
+			/>
 		</>
 	);
 };
