+++ apps/app/src/app/@buyer/feed/FeedListPage/FeedList/Data.tsx
@@ -1,8 +1,10 @@
 import type { MarkSuspense } from "@use-pico/client/type";
 import { Container } from "@use-pico/client/ui/container";
+import { EmptyState } from "@use-pico/client/ui/empty-state";
 import { withFeedQuery } from "@zbav-se.me/sdk/query/buyer/feed";
-import type { FC } from "react";
+import { type FC, useMemo } from "react";
 import { CreateButton } from "./CreateButton";
+import { Empty } from "./Empty";
 import { Item } from "./Item";
 
 export namespace Data {
@@ -15,7 +17,7 @@ export const Data: FC<Data.Props> = ({ _suspense, ...props }) => {
 	/**
 	 * This is intentional to trigger parent suspense
 	 */
-	const { data: feedList } = withFeedQuery.useCollectionQuery({
+	const { data: feedCollection } = withFeedQuery.useCollectionQuery({
 		filter: {
 			type: "user",
 		},
@@ -26,6 +28,21 @@ export const Data: FC<Data.Props> = ({ _suspense, ...props }) => {
 			},
 		],
 	});
+
+	const check = useMemo(() => {
+		return [
+			{
+				check() {
+					return !feedCollection.length;
+				},
+				render() {
+					return <Empty />;
+				},
+			},
+		] satisfies EmptyState.Check[];
+	}, [
+		feedCollection,
+	]);
 
 	return (
 		<Container
@@ -39,16 +56,18 @@ export const Data: FC<Data.Props> = ({ _suspense, ...props }) => {
 			}}
 			{...props}
 		>
-			{feedList.map((feedId) => {
-				return (
-					<Item
-						key={feedId}
-						feedId={feedId}
-					/>
-				);
-			})}
+			<EmptyState check={check}>
+				{feedCollection.map((feedId) => {
+					return (
+						<Item
+							key={feedId}
+							feedId={feedId}
+						/>
+					);
+				})}
 
-			<CreateButton />
+				<CreateButton />
+			</EmptyState>
 		</Container>
 	);
 };
