+++ apps/app/src/app/@buyer/feed/FeedListPage/FeedListPage.tsx
@@ -1,7 +1,6 @@
 import { translator } from "@use-pico/common/translator";
 import { TitleContainer } from "@zbav-se.me/ui/container";
 import type { FC } from "react";
-import { FEED_LIMIT } from "~/app/@common/limit/Limit";
 import { BackHomeButton } from "~/app/@common/nav/BackHomeButton";
 import { HomeMenuButton } from "~/app/@user/home/~public/HomeMenuButton";
 import { FeedList } from "./FeedList";
@@ -31,10 +30,7 @@ export const FeedListPage: FC<FeedListPage.Props> = ({ ui, ...props }) => {
 			right={<HomeMenuButton />}
 			{...props}
 		>
-			<FeedList
-				data-ui={"FeedList-[FeedList]"}
-				limit={FEED_LIMIT}
-			/>
+			<FeedList data-ui={"FeedListPage-[FeedList]"} />
 		</TitleContainer>
 	);
 };
